requirements.txt
asyncio
python-dotenv>=0.20
telethon>=1.24.0